chore: add default option extentions
We can extend options passed to the core-api by defining an extension
object, this is so we can pass http options to the http api client
but still get type safety.

We define a default extension type of `{}` so we don't need to define
the extension in places we aren't extending it (e.g. ipfs-core).

Problem is, JSDoc doesn't pick up the default extension object so we
need to add it to the function definitions.

Refs: https://github.com/ipfs/js-ipfs/pull/3917#discussion_r761886258