fix: replace slice with subarray for increased performance
In several places we call `.slice` as a way to transform `BufferList`s
to `Uint8Array`s. Due to refactors in some places we are now calling
`.slice` on `Uint8Array`s which is a memory-copy operation.

In other places `Uint8ArrayList`s are now returned instead of `BufferList`s
on which `.slice` is also a memory-copy operation.

Swap `.slice` for `.subarray` which is no-copy for `Uint8Array`s and
can be no-copy for `Uint8ArrayList`s too, where there is only a single
backing buffer.

In places where we need to transform multiple `Uint8ArrayList`s to
multiple `Uint8Array`s, yield the iterators of the `Uint8ArrayList`s
as this is also a no-copy operation.